refactor: rewrite file to ECMAScript notation
